Dashboard: migrating AZ's new optimization
