+++ web/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "dependencies": {
     "@ant-design/icons": "^4.7.0",
-    "@headlessui/react": "^1.7.3",
+    "@headlessui/react": "^1.7.11",
     "@heroicons/react": "^2.0.12",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
