--- web/src/components/AZ's LogIn/TestingLogIn.page.tsx
@@ -1,20 +1,0 @@
-//  Sign Up Page for Dashboard
-//      by AZ
-
-//  TBC:
-//      [] Component
-//      [] Cognito
-//      [] autocomplete
-//      [] onEnter
-
-
-
-const TestingLogIn = () => {
-    return (
-        <div>
-            <div></div>
-        </div>
-    )
-}
-
-export default TestingLogIn
