+++ web/src/components/Dashboard/Dashboard.css
@@ -1,4 +1,4 @@
-@import "~antd/dist/antd.css";
+/* @import "~antd/dist/antd.css";
 
 #main {
   background-color: #333;
@@ -57,4 +57,4 @@
   min-height: 100vh;
   grid-template-columns: repeat(3, 1fr);
   grid-gap: 32px;
-}
+} */
