+++ web/src/components/Dashboard/Dashboard.scss
@@ -1,9 +1,6 @@
 // @import "~antd/dist/antd.css";
 
 #main {
-    background-color: #333;
-
-
     #searchBar {
         display: grid;
         margin: 20px auto;
@@ -26,25 +23,4 @@
         }
     }
 
-    #filterBar {
-        display: grid;
-        position: relative;
-        grid-template-columns: repeat(4, 1fr);
-        zoom: 100%;
-        margin: 20px auto;
-        width: fit-content;
-        text-align: center;
-    }
-
-    #main-container {
-        display: grid;
-        position: relative;
-        max-width: 800px;
-        width: 90vw;
-        margin: 20px auto;
-        min-height: 100vh;
-        grid-template-columns: repeat(3, 1fr);
-        grid-auto-rows: min-content;
-        grid-gap: 32px;
-    }
 }
