+++ web/src/index.css
@@ -15,3 +15,7 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
 }
+
+body * {
+    box-sizing: border-box;
+}
